% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{animals}
\alias{animals}
\title{Dataset of archaeological measurements with terminus post and ante quem dating}
\format{
A data frame with 5533 rows and 6 variables:
\itemize{
\item taxa 4 species names (imaginary)
\item site 64 site names (abbreviated)
\item us 1013 us (abbreviated), within these sites
\item tpq terminal post quem, ranging from -225 to 375
\item taq terminal ante quem, ranging from -175 to 700
\item value the parameter of interest measured
}
}
\source{
Unpublished dataset, modified by the authors to become a toy dataset.
}
\usage{
animals
}
\description{
This is a real but "anonymized" dataset of archaeological data.
}
\details{
The original dataset aimed at measured the evolution of the \code{value} index,
of four \code{taxa}, on remains recovered in \code{us} within \code{site} wand with datations
provided as \code{tpq} and \code{taq}.
}
\keyword{datasets}
