% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\docType{data}
\name{animals100}
\alias{animals100}
\title{Animals with 100 permutations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12000 rows and 4 columns.
}
\usage{
animals100
}
\description{
See \link{animals}. This is obtained with:\preformatted{set.seed(2329)
animals100 <-  quake(animals, tpq, taq, value, k=100, group=taxa)
}
}
\keyword{datasets}
