% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{comparison_testers}
\alias{comparison_testers}
\alias{wilcox_p}
\alias{kruskal_p}
\alias{aov_p}
\title{Comparisons testers}
\usage{
wilcox_p(df)

kruskal_p(df)

aov_p(df)
}
\arguments{
\item{df}{\code{\link[=tibble]{tibble()}} with two columns named \code{g} and \code{y},
and with only two levels in \code{group}.}
}
\value{
numeric p.value (or \code{NA_real} if not enough data or groups)
}
\description{
These testers are used internally by \href{testers}{test} but may be of interest outside.
}
\details{
These testers are thin wrappers around \link[stats:wilcox.test]{stats::wilcox.test}, \link[stats:kruskal.test]{stats::kruskal.test}
and \link[stats:aov]{stats::aov}. Type \code{wilcox_p} (no bracket) for how to build your own.
}
\section{Functions}{
\itemize{
\item \code{wilcox_p}: wilcoxon's rank test for pairwise differences

\item \code{kruskal_p}: kruskal's test for global differences

\item \code{aov_p}: analysis of variance test for both global and pairwise differences
}}

\examples{
# dummy tibble
x <- tibble::tibble(g=rep(letters[1:2], each=10),
                   y=stats::runif(20, -1, 1))

wilcox_p(x)
aov_p(x)

# on >2 groups
x$g[15:20] <- "c"
kruskal_p(x)
aov_p(x)
}
