% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_g}
\alias{df_g}
\title{Toy dataset with centered/interval dating}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 5 columns.
}
\usage{
df_g
}
\description{
Only used for examples
}
\details{
This is obtained with:\preformatted{set.seed(2329)
df_g <- tibble::tibble(c14  = round(runif(8, -100, 100)), # best prediction
                       sd1   = round(runif(8, 5, 20)),     # one sd
                       species = rep(c("fox", "hound"), each=4), # dummy grouping
                        site = rep(letters[1:2], each=2, times=2), # dummy sites
                       mes  = c(1:4, 4:1) + runif(8, -0.5, 0.5))
usethis::use_data(df_g, overwrite=TRUE)
}
}
\keyword{datasets}
