% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{df_u}
\alias{df_u}
\title{Toy dataset with terminus ante and post quem dating}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 5 columns.
}
\usage{
df_u
}
\description{
Only used for examples
}
\details{
This is obtained with:\preformatted{}
}
\keyword{datasets}
