% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diachrony.R
\name{diachrony_uniform}
\alias{diachrony_uniform}
\title{Prepare a data.frame for permutations}
\usage{
diachrony_uniform(df, tpq, taq, y)
}
\arguments{
\item{df}{\code{\link[=data.frame]{data.frame()}} to work on}

\item{tpq, taq}{colname to use for tpq and taq (when shaking_gaussian)}

\item{y}{colname to use for the value of interest}
}
\description{
This allows to ease and fasten permutations because the raw data is
\code{prepare}d once for all. This is a thin wrapper around \code{\link[dplyr:rename]{dplyr::rename()}} that
also checks conditions before permutations.
}
\examples{
# let's say animals had different names
x <- animals \%>\% dplyr::rename(post=tpq, ante=taq, lsi=value)
x
x \%>\% diachrony_uniform(post, ante, lsi)
}
