% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pataqu-package.R
\docType{package}
\name{pataqu-package}
\alias{pataqu}
\alias{pataqu-package}
\title{pataqu: Permutationnal Analysis of Terminus Ante and post QUem data}
\description{
pataqu uses permutationnal approach to deal with data that comes with uncertainties on x. It it typically aimed at working with temporal data that are not exactly defined but comes with lower (and upper) bounds, also called terminus ante (and post) quem. It can also be used when dating, typically radiocarbon datation, comes with a confidence interval likely to affect interpretation. More generally, it is a permutationnal approach that allows to inspect both graphically and statistically the effects of x uncertainties.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vbonhomme/pataqu}
  \item \url{https://vbonhomme.github.io/pataqu/}
  \item Report bugs at \url{https://github.com/vbonhomme/pataqu/issues}
}

}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{bonhomme.vincent@gmail.com} (\href{https://orcid.org/0000-0002-2742-6349}{ORCID})

Authors:
\itemize{
  \item Allowen Evin (\href{https://orcid.org/0000-0003-4515-1649}{ORCID})
}

}
\keyword{internal}
