% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\alias{predictor_loess}
\title{Predictor function to summarise results after shaking}
\usage{
predictor_loess(df, x_prediction, ...)
}
\arguments{
\item{df}{\code{\link[=tibble]{tibble()}}, typically after \link{shake}. Must have a .x and and .y columns.}

\item{x_prediction}{numeric vector on x values on which to predict}

\item{...}{additional parameters for each predictors (see Details)}
}
\description{
These functions are used internally by \code{diachrony_} and \code{synchrony_} functions,
but they may be of interest for custom constructions. See Details.
}
\details{
In \code{df}, \code{x_new} and \code{y} columns must be present. This ease and fasten permutations
because the raw data is \code{prepare}d once for all.

Most of these functions are thin wrappers around built-in  predictors such as
\code{\link[stats:loess]{stats::loess()}}, \code{\link[stats:lm]{stats::lm()}}, \code{\link[mgcv:gam]{mgcv::gam()}}, etc.

These predictors follow these steps:
\enumerate{
\item fit the \code{y} (value of interest) using the (typically shaked) \code{x}. We thus obtain
a continuous model that we use to:
\item predict new \code{y} values on a range of \code{x}. These \code{x} values are typically fixed, so that
we can later obtain confidence intervals, etc. for each of these points.
}

They thus aim at summarising new \code{y} values at fixed \code{x} values, using

\code{predict_bin} makes no intermediate
adjustment and summarises raw value either using mean or median (see Examples)
}
\section{Functions}{
\itemize{
\item \code{predictor_loess}: predict using loess
}}

\examples{

# replicability
set.seed(2329)
x_pred <- 1:100
df <- data.frame(x_new=x_pred, y=runif(100, -1, 1) + x_pred/100)
x_pred <- 1:100
pred_075 <- predictor_loess(df, x_pred)
pred_05 <- predictor_loess(df, x_pred, span=0.5)
plot(df$x_new, df$y)
lines(pred_075$x_new, pred_075$y, col="red")
lines(pred_05$x_new, pred_05$y, col="blue")

}
