% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{quake}
\alias{quake}
\title{Generate new datasets with new values}
\usage{
quake(x, k = 1, shaker = shake_uniform, ...)
}
\arguments{
\item{x}{\code{\link[=tibble]{tibble()}}}

\item{k}{number of new datasets to create}

\item{shaker}{one of the \href{shake()}{shaker} functions (\code{\link[=shake_uniform]{shake_uniform()}} by default)}

\item{...}{arguments expected by the selected \href{shake()}{shaker} function. They
must be provided and named. See examples}
}
\value{
a 'shaken' \code{\link[=tibble]{tibble()}} with new columns and iteration index


}
\description{
This is a wrapper around \href{shake()}{shakers} that allows to run it \code{k} times
and adds and index to the resulting tibble. As the typical entry point in permutationnal
analysis it also checks your data to avoid further problems.
}
\examples{
set.seed(2329) # replicability
# shaking uniform
df_u
df_u \%>\% quake(k=5, shake_uniform, min=tpq, max=taq)
# not that you can omit shakers' argument names, providing they come in the right order
# eg:  df_u \%>\% quake(k=5, shake_uniform, tpq, taq)

# shaking gaussian
df_g
df_g \%>\% quake(k=5, shake_gaussian, mean=c14, sd=sd1)

# on a more realistic dataset bigger baby:
# animals \%>\% quake(k=2, shake_uniform, tpq, taq)
}
