% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quake.R
\name{quake_uniform}
\alias{quake_uniform}
\alias{quake}
\title{Use permutations to visualize and test uncertainties on x}
\usage{
quake_uniform(
  df,
  tpq,
  taq,
  y,
  group,
  k = 10,
  predictor_fun = predictor_loess,
  x_prediction = 30,
  ...
)
}
\arguments{
\item{df}{\code{\link[=data.frame]{data.frame()}} to work on}

\item{tpq, taq}{colname to use for tpq and taq (when shaking_gaussian)}

\item{y}{colname to use for the value of interest}

\item{group}{colname (optionnal) whether you have groups}

\item{k}{number of permutations (10 by default)}

\item{predictor_fun}{function one of \link{predictor}}

\item{x_prediction}{on which to predict new values. See Details.}

\item{...}{additional arguments to \code{predictor_fun}}
}
\description{
More details here.
}
\details{
If \code{x_prediction} can be passed as a single numeric or as
a vector of numeric. If a single numeric is passed, a regular sequence of this length
will be created between \code{min(tpq)} and \code{max(taq)}, ie covering the full temporal span.
If a vector of numeric is passed, then these values will be used. By default, it
eis \code{30}, so that 30 dates covering the temporal span will be used. See examples.
}
\examples{
set.seed(2329) # for replicability

# let's say animals had different names
x <- animals \%>\% dplyr::rename(post=tpq, ante=taq, lsi=value)
x
x \%>\% quake_uniform(post, ante, lsi)

animals \%>\% quake_uniform(tpq, taq, value, k=2, group=taxa)

\dontrun{
library(ggplot2)
x \%>\% ggplot() +
  aes(x_new, y, group=k, col=group) +
  geom_point(size=0.3) +
  theme_minimal() +
 geom_smooth(aes(group=NULL, col=group))
}
}
