% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\name{synchrony_testers}
\alias{synchrony_testers}
\alias{wilcox_p}
\alias{kruskal_p}
\alias{aov_p}
\title{Pairwise comparisons testers}
\usage{
wilcox_p(x)

kruskal_p(x)

aov_p(x)
}
\arguments{
\item{x}{\code{\link[=tibble]{tibble()}} with two columns named \code{group} and \code{y},
and with only two levels in \code{group}.}
}
\value{
numeric p.value (or \code{NA_real} if not enough data or groups)
}
\description{
These testers return p-value for \code{\link[=synchrony]{synchrony()}} pairwise comparisons.
}
\details{
These testers are thin wrappers around \link[stats:wilcox.test]{stats::wilcox.test}, \link[stats:kruskal.test]{stats::kruskal.test}
and \link[stats:aov]{stats::aov}. Type \code{wilcox_p} (no bracket) for how to build your own.
}
\section{Functions}{
\itemize{
\item \code{wilcox_p}: wilcoxon's rank test for pairwise differences

\item \code{kruskal_p}: kruskal's test for global differences

\item \code{aov_p}: analysis of variance test for both global and pairwise differences
}}

\examples{
# dummy tibble
x <- tibble::tibble(group=rep(letters[1:2], each=10),
                   y=stats::runif(20, -1, 1))

wilcox_p(x)
aov_p(x)

# on >2 groups
x$group[15:20] <- "c"
kruskal_p(x)
aov_p(x)
}
