% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin}
\usage{
bin(
  df,
  y = y,
  fun = stats::median,
  by = NULL,
  x_pred = cutter_number(df, 30, x_new),
  k = k
)
}
\arguments{
\item{df}{\code{\link[=tibble]{tibble()}} typically returned by \code{\link[=quake]{quake()}}}

\item{y}{colname of the value of interest. Default to \code{y}.}

\item{fun}{to use to summarise \code{y}}

\item{by}{colname for grouping structure (besides \code{k}). Default to \code{NULL}}

\item{x_pred}{the sequence of x for which to predict new values after fitting}

\item{k}{colname for iteration . Default to \code{k}.}
}
\description{
Bin and summarise without \link{fitting}
}
