% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{bin}
\alias{bin}
\title{Bin}
\usage{
bin(df, y = y, x = x_new, by = NULL, fun = stats::median, k = k)
}
\arguments{
\item{df}{\code{\link[=tibble]{tibble()}} typically returned by \code{\link[=quake]{quake()}}}

\item{y, x}{colnames to use. Default to \code{y}/\code{x_new}}

\item{by}{colname for grouping structure (besides \code{k}). Default to \code{NULL}}

\item{fun}{to use to summarise \code{y}}

\item{k}{colname for iteration . Default to \code{k}.}
}
\description{
Bin and summarise without \link{fitting}
}
\examples{
animals_q \%>\% bin(y=value, by=taxa, fun=mean)

animals_q \%>\% bin(y=value, fun=mean)
}
