% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shake.R
\name{shake}
\alias{shake}
\alias{shake_uniform}
\alias{shake_gaussian}
\title{Generate new x values}
\usage{
shake_uniform(x, min_col, max_col)

shake_gaussian(x, mean_col, sd_col)
}
\arguments{
\item{x}{\code{\link[=tibble]{tibble()}}}

\item{min_col}{colname to use for taq}

\item{max_col}{colname to use for tpq}

\item{mean_col}{colname to use for mean}

\item{sd_col}{colname to use for sd}
}
\value{
a 'shaked'  \code{\link[=tibble]{tibble()}}
}
\description{
Used for each permutation to generate new x values.
Different shakers are available depending on the type of
uncertainties you have
}
\section{Functions}{
\itemize{
\item \code{shake_uniform}: uniform distribution

\item \code{shake_gaussian}: gaussian distribution
}}

\examples{

animals \%>\% shake_uniform(taq, tpq)

animals \%>\% dplyr::mutate(mean=0, sd=50) \%>\% shake_gaussian(mean, sd)

}
