% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shake.R
\name{shake}
\alias{shake}
\alias{shake_uniform}
\alias{shake_gaussian}
\title{Generate new x values}
\usage{
shake_uniform(x)

shake_gaussian(x)
}
\arguments{
\item{x}{\code{\link[=tibble]{tibble()}} typically obtained with \code{\link[=prepare]{prepare()}}}
}
\value{
a 'shaked'  \code{\link[=tibble]{tibble()}}
}
\description{
Used internally for each permutation to generate new x values.
Different shakers are available depending on the type of
uncertainties you have.
}
\details{
expects \code{tpq}, \code{taq} colnames for \code{shake_uniform},
and \code{mean}, \code{sd} for \code{shake_gaussian}
}
\section{Functions}{
\itemize{
\item \code{shake_uniform}: uniform distribution

\item \code{shake_gaussian}: gaussian distribution
}}

\examples{
set.seed(2329) # replicability
tibble::tibble(tpq=c(-50, 100), taq=c(50, 200)) \%>\% shake_uniform()
tibble::tibble(mean=c(0, 100), sd=c(5, 20)) \%>\% shake_gaussian()

}
