% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti.R
\name{spaghetti}
\alias{spaghetti}
\title{Spaghetti plot after permutations}
\usage{
spaghetti(x)
}
\arguments{
\item{x}{the result of a permutation function}
}
\description{
ggplot2 primer for displaying the results of permutations
}
\examples{
x_loess <- animals \%>\%
  diachrony_uniform(tpq, taq, value, k=5,
  group=taxa, predictor_fun=predictor_loess)

spaghetti(x_loess)

# with a lm now
x_lm <- animals \%>\%
  diachrony_uniform(tpq, taq, value, k=5,
  group=taxa, predictor_fun=predictor_lm)

spaghetti(x_lm)
}
