% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaghetti.R
\name{spaghetti}
\alias{spaghetti}
\title{Spaghetti plot after permutations}
\usage{
spaghetti(x, ...)
}
\arguments{
\item{x}{the result of a permutation function}

\item{...}{additional parameters to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}
}
\description{
ggplot2 primer on the result of \link{quake} for displaying the results of permutations.
This allows inspecting the effect of \link{predictor}s.
}
\examples{
animals_q \%>\%
spaghetti()

# you can easily customize it
# color palette from https://www.colourlovers.com/palette/1473/Ocean_Five
colors <- c("bird"="#00A0B0","cat"="#CC333F", "frog"="#CBE86B", "mouse"="#EDC951")
library(ggplot2)
animals_q \%>\%
spaghetti(alpha=0.5, size=0.1) +
  scale_color_manual(values=colors) +
  xlab("year") + ylab("value of interest") +
  guides(colour=guide_legend("taxa", override.aes=list(alpha=1, size=2)))
}
